+++ src/util.ts
@@ -1,4 +1,4 @@
-import path from 'node:path'
+import path from 'path'
 import * as vscode from 'vscode'
 
 export const PACKAGE_NAME = 'saveEditorLayout'
@@ -9,6 +9,27 @@ export function readConfig() {
     config = vscode.workspace.getConfiguration(PACKAGE_NAME)
 }
 
-export function getFileName(filePath) {
-    return filePath ? path.parse(filePath).base : ''
+export function getList() {
+    return config.get('list', [])
+}
+
+export async function updateList(data) {
+    await config.update('list', data, vscode.ConfigurationTarget.Workspace)
+}
+
+export function getSnapshotName(snapshotName) {
+    const list = getList()
+    const item = list.find((item) => item.file === snapshotName)
+
+    return item ? item.name : undefined
+}
+
+export function getNewFileFullPath(dbPath, file) {
+    return path.join(path.dirname(dbPath), file)
+}
+
+export function showMsg(msg, error = false, args: string[] = []) {
+    return error
+        ? vscode.window.showErrorMessage(`Save Editors Layout: ${msg}`, ...args)
+        : vscode.window.showInformationMessage(`Save Editors Layout: ${msg}`, ...args)
 }